avoid a variety of warnings for MSVC
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@21684 f0d54148-0727-0410-94bb-9a71ac55c965

+++ ogr/ogrsf_frmts/bna/ogrbnadatasource.cpp
@@ -204,7 +204,7 @@ int OGRBNADataSource::Open( const char * pszFilename, int bUpdateIn)
 
         while(1)
         {
-            int offset = VSIFTellL(fp);
+            int offset = (int) VSIFTellL(fp);
             int line = curLine;
             record =  BNA_GetNextRecord(fp, &ok, &curLine, FALSE, BNA_READ_NONE);
             if (ok == FALSE)
