+++ ogr/ogrsf_frmts/kml/ogrkmldatasource.cpp
@@ -297,7 +297,7 @@ int OGRKMLDataSource::Create( const char* pszName, char** papszOptions )
 /* -------------------------------------------------------------------- */
     VSIFPrintfL( fpOutput_, "<?xml version=\"1.0\" encoding=\"utf-8\" ?>\n" );	
 
-    nSchemaInsertLocation_ = VSIFTellL( fpOutput_ );
+    nSchemaInsertLocation_ = (int) VSIFTellL( fpOutput_ );
     
     VSIFPrintfL( fpOutput_, "<kml xmlns=\"http://www.opengis.net/kml/2.2\">\n<Document>" );
